new/DPL/requirements.txt
# LLM & LangChain (совместимые версии 0.3.x)
# Оставляем для обратной совместимости, но основной клиент - LiteLLM
langchain>=0.3.0,<1.0.0
langchain-openai>=0.2.0,<1.0.0
langchain-core>=0.3.0,<1.0.0

# LiteLLM - единый интерфейс для 100+ LLM
litellm>=1.0.0

# Data validation
pydantic>=2.9.0
pydantic-settings>=2.5.0

# Environment
python-dotenv>=1.0.0

# HTTP requests
requests>=2.31.0

# Tokenization
tiktoken>=0.7.0

# Progress bars
tqdm>=4.66.0

# Data processing
numpy>=1.26.0

# Visualization
matplotlib>=3.9.0

# === YouTube & Transcription ===

# YouTube download
yt-dlp>=2024.8.0

# Whisper transcription (Hugging Face)
transformers>=4.44.0
accelerate>=0.33.0

# PyTorch
torch>=2.1.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0
